Fix nonstandard, unsigned OpenMP loop counter
OpenMP does not yet permit unsigned loop counters. Most compilers seem
to allow them anyway, but a few have been seen to complain.

Mistake introduced in:
    commit 82ed0fde32d5e85b238e821af1e542574d8e9fe7
    Author: Chris Pavlina <pavlina.chris@gmail.com>
    Date:   Sun Aug 28 02:05:49 2016 -0400

    Fix shared data access in raytracer, tidy up render loop

+++ 3d-viewer/3d_rendering/3d_render_raytracing/c3d_render_raytracing.h
@@ -40,7 +40,6 @@
 #include <plugins/3dapi/c3dmodel.h>
 
 #include <map>
-#include <cstddef>
 
 /// Vector of materials
 typedef std::vector< CBLINN_PHONG_MATERIAL > MODEL_MATERIALS;
@@ -111,7 +110,7 @@ private:
     unsigned long int m_stats_start_rendering_time;
 
     /// Save the number of blocks progress of the render
-    size_t m_nrBlocksRenderProgress;
+    long m_nrBlocksRenderProgress;
 
     CPOSTSHADER_SSAO m_postshader_ssao;
